Remove failing example from androidx-samples module
