+++ koin-projects/examples/androidx-samples/src/main/java/org/koin/sample/androidx/MainApplication.kt
@@ -22,7 +22,6 @@ class MainApplication : Application() {
                     mvpModule,
                     mvvmModule,
                     scopeModule,
-                    failingScopeModule,
                     passingScopeModule,
                     objectScopeModule
             ))
