+++ koin-projects/examples/androidx-samples/src/main/java/org/koin/sample/androidx/components/objectscope/Consumer.kt
@@ -1,0 +1,3 @@
+package org.koin.sample.androidx.components.objectscope
+
+class Consumer(val interceptors: List<Interceptor>, val someService: SomeService)
