+++ koin-projects/examples/androidx-samples/src/main/java/org/koin/sample/androidx/components/objectscope/Service.kt
@@ -1,0 +1,6 @@
+package org.koin.sample.androidx.components.objectscope
+
+import androidx.fragment.app.FragmentActivity
+
+class SomeService(val activity: FragmentActivity)
+class OtherService(val fragment: FragmentActivity)
